Refactored the rest of the pages since the scale of the changes were quite small now that I went through the first couple phases.
