+++ src/app/actions/tags.ts
@@ -2,32 +2,11 @@
 
 import { db } from '@/lib/db';
 import { tags, posts, postTags } from '@/lib/db/schema';
-import { revalidatePath } from 'next/cache';
 import { desc, count, sql, eq } from 'drizzle-orm';
-
-import * as mediaRepo from '@/lib/db/repositories/media';
 import * as postsRepo from '@/lib/db/repositories/posts';
 
 export async function getPostTags(postId: number) {
     return postsRepo.getPostTags(postId);
-}
-
-export async function getMediaItems(filters?: { search?: string; sortBy?: string; limit?: number; cursor?: string; }) {
-    return mediaRepo.getMediaItems(filters);
-}
-
-export type { TimelinePost } from '@/lib/db/repositories/posts';
-
-export async function getTimelinePosts(filters?: { search?: string; sortBy?: string; limit?: number; cursor?: string; }) {
-    return postsRepo.getTimelinePosts(filters);
-}
-
-export async function deleteMediaItems(ids: number[], deleteFiles: boolean) {
-    const result = await mediaRepo.deleteMediaItems(ids, deleteFiles);
-    revalidatePath('/gallery');
-    revalidatePath('/timeline');
-    revalidatePath('/');
-    return result;
 }
 
 export async function getTopTags(sort: 'count' | 'new' | 'recent' = 'count') {
