+++ src/app/actions/timeline.ts
@@ -1,0 +1,9 @@
+'use server';
+
+import * as postsRepo from '@/lib/db/repositories/posts';
+
+export type { TimelinePost } from '@/lib/db/repositories/posts';
+
+export async function getTimelinePosts(filters?: { search?: string; sortBy?: string; limit?: number; cursor?: string; }) {
+    return postsRepo.getTimelinePosts(filters);
+}
