--- src/types/gallery.ts
@@ -1,3 +1,0 @@
-export * from './media';
-export * from './posts';
-export * from './users';
